backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
bcrypt
python-jose[cryptography]
python-multipart
google-genai
python-dotenv
httpx
